+++ WebBanThuocBaoVeThucVat/src/main/java/controller/Admin/Dashboard.java
@@ -1,4 +1,4 @@
-package controller;
+package controller.Admin;
 
 import Service.ProductsService;
 import Service.UserService;
@@ -13,7 +13,7 @@ import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.List;
 
-@WebServlet(name = "Dashboard", value = "/admin_dashboard")
+@WebServlet(name = "dashboard", value = "/dashboard")
 public class Dashboard extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
@@ -22,13 +22,6 @@ public class Dashboard extends HttpServlet {
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        List<User>dsUser= UserService.getInstance().getDSUsers();
-        List<Products>dsProduct = ProductsService.getInstance().numOfPro();
-//        System.out.println(dsProduct);
-        int numOfGuest= dsUser.size();
-        int numOfPro=dsProduct.size();
-        req.setAttribute("numUser",numOfGuest);
-        req.setAttribute("numPro",numOfPro);
-        req.getRequestDispatcher("admin_page/admin.jsp").forward(req,resp);
+        req.getRequestDispatcher("admin_page/dashboard.jsp").forward(req,resp);
     }
 }
